AI Chatbot2/requirements.txt
django>=5.1.6
djongo>=1.3.6
python-dotenv>=1.0.0
requests>=2.31.0
tiktoken>=0.5.2
google-generativeai>=0.3.2